+++ prog/dftb+/lib_dftbplus/inputdata.F90
@@ -39,6 +39,7 @@ module dftbp_inputdata_module
 
   public :: control, TGeometry, slater, inputData, XLBOMDInp, TParallelOpts
   public :: TBlacsOpts
+  public :: TRangeSepInp
   public :: init, destruct
 #:if WITH_TRANSPORT
   public :: TNEGFInfo
@@ -75,6 +76,15 @@ module dftbp_inputdata_module
   end type TLbfgsInput
 
 
+  !> Range separation input
+  type TRangeSepInp
+    real(dp) :: screeningThreshold
+    real(dp) :: cutoffRed
+    real(dp) :: omega
+    character(lc) :: rangeSepAlg
+  end type TRangeSepInp
+
+
   !> Main control data for program as extracted by the parser
   type control
 
@@ -88,7 +98,7 @@ module dftbp_inputdata_module
     logical :: tScc        = .false.
 
     !> l-shell resolved SCC
-    logical :: tOrbResolved = .false.
+    logical :: tShellResolved = .false.
 
     !> SCC tolerance
     real(dp) :: sccTol      = 0.0_dp
@@ -440,6 +450,8 @@ module dftbp_inputdata_module
 
     !> LBFGS input
     type(TLbfgsInput), allocatable :: lbfgsInp
+
+    type(TRangeSepInp), allocatable :: rangeSepInp
 
 
   #:if WITH_SOCKETS