+++ prog/modes/modeprojection.F90
@@ -106,7 +106,7 @@ contains
         end if
         vTmp(:) = inertia(:,ii)
         do iAt = 1, nMovedAtom
-          call cross3(rTmp, vTmp, geo%coords(:,iAt) - centreOfMass)
+          rTmp = cross3(vTmp, geo%coords(:,iAt) - centreOfMass)
           vectorsToNull((iAt - 1) * 3 + 1 : iAt * 3, nToNull - ii + 1) = rTmp
         end do
       end do